+++ packages/ses/src-xs/compartment.js
@@ -1,0 +1,284 @@
+/* eslint-disable @endo/no-nullish-coalescing */
+import {
+  Map,
+  Object,
+  SyntaxError,
+  TypeError,
+  WeakMap,
+  arrayMap,
+  create,
+  defineProperty,
+  entries,
+  fromEntries,
+  getOwnPropertyDescriptor,
+  globalThis,
+  mapDelete,
+  mapGet,
+  mapSet,
+  uncurryThis,
+  weakmapGet,
+  weakmapSet,
+} from '../src/commons.js';
+import { nativeGetGlobalThis } from './commons.js';
+import {
+  makeCompartmentConstructor,
+  compartmentOptions,
+} from '../src/compartment.js';
+import { getGlobalIntrinsics } from '../src/intrinsics.js';
+import { tameFunctionToString } from '../src/tame-function-tostring.js';
+import { chooseReporter } from '../src/reporting.js';
+
+const muteReporter = chooseReporter('none');
+
+export const ShimStartCompartment = makeCompartmentConstructor(
+  makeCompartmentConstructor,
+  getGlobalIntrinsics(globalThis, muteReporter),
+  tameFunctionToString(),
+);
+
+export const shimCompartmentPrototype = ShimStartCompartment.prototype;
+
+export const shimEvaluate = uncurryThis(shimCompartmentPrototype.evaluate);
+export const shimImport = uncurryThis(shimCompartmentPrototype.import);
+export const shimImportNow = uncurryThis(shimCompartmentPrototype.importNow);
+
+export const shimGetGlobalThis = uncurryThis(
+  getOwnPropertyDescriptor(ShimStartCompartment.prototype, 'globalThis').get,
+);
+
+export const privateFields = new WeakMap();
+
+const adaptVirtualModuleSource = ({
+  execute,
+  imports = [],
+  exports = [],
+  reexports = [],
+}) => {
+  const resolutions = create(null);
+  let i = 0;
+  return {
+    execute(environment) {
+      const fakeCompartment = {
+        importNow(specifier) {
+          return environment[specifier];
+        },
+      };
+      execute(environment, fakeCompartment, resolutions);
+    },
+    bindings: [
+      ...arrayMap(imports, specifier => {
+        const resolved = `_${i}`;
+        i += 1;
+        resolutions[specifier] = resolved;
+        return { importAllFrom: specifier, as: resolved };
+      }),
+      ...arrayMap(reexports, specifier => ({ exportAllFrom: specifier })),
+      ...arrayMap(exports, name => ({ export: name })),
+    ],
+  };
+};
+
+const adaptModuleSource = source => {
+  if (source.execute) {
+    return adaptVirtualModuleSource(source);
+  }
+  // eslint-disable-next-line no-underscore-dangle
+  if (source.__syncModuleProgram__) {
+    throw new SyntaxError(
+      'XS native compartments do not support precompiled module sources',
+    );
+  }
+  return source;
+};
+
+const adaptModuleDescriptor = (
+  descriptor,
+  specifier,
+  compartment = undefined,
+) => {
+  if (Object(descriptor) !== descriptor) {
+    throw new TypeError('module descriptor must be an object');
+  }
+  if (descriptor.namespace !== undefined) {
+    return descriptor;
+  }
+  if (descriptor.source !== undefined) {
+    return {
+      source: adaptModuleSource(descriptor.source),
+      importMeta: descriptor.importMeta,
+      specifier: descriptor.specifier,
+    };
+  }
+  // Legacy support for record descriptors.
+  if (descriptor.record !== undefined) {
+    if (
+      descriptor.specifier === specifier ||
+      descriptor.specifier === undefined
+    ) {
+      return {
+        source: adaptModuleSource(descriptor.record),
+        specifier,
+        importMeta: descriptor.importMeta,
+      };
+    } else {
+      if (compartment === undefined) {
+        throw new TypeError(
+          'Cannot construct forward reference module descriptor in module map',
+        );
+      }
+      const { descriptors } = weakmapGet(privateFields, compartment);
+      mapSet(descriptors, descriptor.specifier, {
+        compartment,
+        namespace: specifier,
+      });
+      return {
+        source: adaptModuleSource(descriptor.record),
+        specifier: descriptor.specifier,
+        importMeta: descriptor.importMeta,
+      };
+    }
+  }
+  if (descriptor.specifier !== undefined) {
+    return {
+      namespace: descriptor.specifier,
+      compartment: descriptor.compartment,
+    };
+  }
+  // Legacy support for a source in the place of a descriptor.
+  return { source: adaptModuleSource(descriptor) };
+};
+
+export const adaptCompartmentConstructors = (
+  NativeCompartment,
+  ShimCompartment,
+  maybeHarden,
+) => {
+  function Compartment(...args) {
+    const options = compartmentOptions(...args);
+
+    const {
+      name = undefined,
+      globals = {},
+      transforms = [],
+      resolveHook = () => {
+        throw new TypeError('Compartment requires a resolveHook');
+      },
+      loadHook = undefined,
+      loadNowHook = undefined,
+      importHook = loadHook,
+      importNowHook = loadNowHook,
+      moduleMapHook = () => {},
+      __native__: delegateNative = false,
+      __noNamespaceBox__: noNamespaceBox = false,
+    } = options;
+
+    const modules = delegateNative
+      ? fromEntries(
+          arrayMap(
+            entries(options.modules ?? {}),
+            ([specifier, descriptor]) => [
+              specifier,
+              adaptModuleDescriptor(descriptor, specifier, undefined),
+            ],
+          ),
+        )
+      : {};
+
+    // side table for references from one descriptor to another
+    const descriptors = new Map();
+
+    let nativeOptions = { globals, modules };
+
+    if (importHook) {
+      /** @param {string} specifier */
+      const nativeImportHook = async specifier => {
+        await null;
+        let descriptor =
+          mapGet(descriptors, specifier) ??
+          moduleMapHook(specifier) ??
+          (await importHook(specifier));
+        mapDelete(descriptors, specifier);
+        // eslint-disable-next-line no-use-before-define
+        descriptor = adaptModuleDescriptor(descriptor, specifier, compartment);
+        return descriptor;
+      };
+      nativeOptions = {
+        ...nativeOptions,
+        resolveHook,
+        importHook: nativeImportHook,
+        loadHook: nativeImportHook,
+      };
+    }
+
+    if (importNowHook) {
+      /** @param {string} specifier */
+      const nativeImportNowHook = specifier => {
+        let descriptor =
+          mapGet(descriptors, specifier) ??
+          moduleMapHook(specifier) ??
+          importNowHook(specifier);
+        mapDelete(descriptors, specifier);
+        // eslint-disable-next-line no-use-before-define
+        descriptor = adaptModuleDescriptor(descriptor, specifier, compartment);
+        return descriptor;
+      };
+      nativeOptions = {
+        ...nativeOptions,
+        resolveHook,
+        importNowHook: nativeImportNowHook,
+        loadNowHook: nativeImportNowHook,
+      };
+    }
+
+    const compartment = new NativeCompartment(nativeOptions);
+
+    const nativeGlobalThis = nativeGetGlobalThis(compartment);
+
+    const nativeEval = nativeGlobalThis.eval;
+
+    weakmapSet(privateFields, compartment, {
+      name,
+      transforms,
+      delegateNative,
+      noNamespaceBox,
+      nativeEval,
+      descriptors,
+    });
+
+    const shimOptions = {
+      ...options,
+      __noNamespaceBox__: true,
+      __options__: true,
+    };
+
+    uncurryThis(ShimCompartment)(compartment, shimOptions);
+
+    const shimGlobalThis = shimGetGlobalThis(compartment);
+
+    const ChildCompartment = adaptCompartmentConstructors(
+      nativeGlobalThis.Compartment,
+      shimGlobalThis.Compartment,
+      maybeHarden,
+    );
+
+    defineProperty(nativeGlobalThis, 'Compartment', {
+      value: ChildCompartment,
+      writable: true,
+      configurable: true,
+      enumerable: false,
+    });
+
+    defineProperty(shimGlobalThis, 'Compartment', {
+      value: ChildCompartment,
+      writable: true,
+      configurable: true,
+      enumerable: false,
+    });
+
+    maybeHarden(compartment);
+    return compartment;
+  }
+
+  maybeHarden(Compartment);
+  return Compartment;
+};