+++ packages/ses/src-xs/lockdown-shim.js
@@ -1,0 +1,26 @@
+/**
+ * @module Alters the XS implementation of Lockdown to be backward compatible
+ * with SES, providing Compartment constructors in every Compartment that can
+ * be used with either native ModuleSources or module sources pre-compiled for
+ * the SES Compartment, depending on the __native__ Compartment constructor
+ * option.
+ */
+import { globalThis } from '../src/commons.js';
+import { NativeStartCompartment } from './commons.js';
+import { repairIntrinsics } from '../src/lockdown.js';
+import {
+  ShimStartCompartment,
+  adaptCompartmentConstructors,
+} from './compartment.js';
+
+globalThis.lockdown = options => {
+  const hardenIntrinsics = repairIntrinsics(options);
+  hardenIntrinsics();
+  // Replace global Compartment with a version that is hardened and hardens
+  // transitive child Compartment.
+  globalThis.Compartment = adaptCompartmentConstructors(
+    NativeStartCompartment,
+    ShimStartCompartment,
+    harden,
+  );
+};